added rudimental servicefactory
This servicefactory is reading services and their functions by their simple presence in the right place.
Places are defined as "/src/services/<name>" and in there in the folder "methods/".
Excluded from parsing file beginning with '.' and any unknown subfolders. If you want do disable a method within a service moving to a "disabled"-folder is recommended.